+++ crates/iam-auth/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "iam-auth"
+version.workspace = true
+rust-version.workspace = true
+edition.workspace = true
+authors.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+
+[dependencies]
+# External dependencies
+anyhow = { workspace = true }
+aws-config = { workspace = true }
+aws-sdk-sts = { workspace = true }
+jsonwebtoken = { workspace = true }
+reqwest = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+tokio = { workspace = true }
+zeroize = { workspace = true, features = ["derive"] }
+
+# Internal dependencies
+doc = { path = "../doc" }
+
+[dev-dependencies]
+insta = { workspace = true }
