+++ crates/runtime/src/lib.rs
@@ -7,7 +7,6 @@ mod combine;
 mod container;
 mod derive;
 pub mod harness;
-mod iam_auth;
 mod image_connector;
 mod local_connector;
 mod materialize;
