+++ go/bindings/.snapshots/TestBuildCatalog-one-materialization
@@ -1,1 +1,1 @@
-(*flow.MaterializationSpec)(name:"example/materialization" connector_type:IMAGE config_json:"{\"image\":\"ghcr.io/estuary/materialize-sqlite:dev\",\"config\":{}}" bindings:<resource_config_json:"{\"table\":\"a_table\"}" resource_path:"a_table" collection:<name:"a/derivation" write_schema_json:"{\"$id\":\"file:///build.flow.yaml?ptr=/collections/a~1derivation/schema\",\"properties\":{\"a_key\":{\"type\":\"string\"}},\"required\":[\"a_key\"],\"type\":\"object\"}" key:"/a_key" uuid_ptr:"/_meta/uuid" projections:<ptr:"/_meta/flow_truncated" field:"_meta/flow_truncated" inference:<types:"boolean" title:"Flow truncation indicator" description:"Indicates whether any of the materialized values for this row have been truncated to make them fit inside the limitations of the destination system." exists:MUST > > projections:<ptr:"/a_key" field:"a_key" is_primary_key:true inference:<types:"string" string:<> exists:MUST > > projections:<field:"flow_document" inference:<types:"object" exists:MUST > > projections:<ptr:"/_meta/uuid" field:"flow_published_at" inference:<types:"string" string:<format:"date-time" content_encoding:"uuid" > title:"Flow Publication Time" description:"Flow publication date-time of this document" exists:MUST > > ack_template_json:"{\"_meta\":{\"ack\":true,\"uuid\":\"DocUUIDPlaceholder-329Bb50aa48EAa9ef\"}}" partition_template:<name:"a/derivation/ffffffffffffffff" replication:3 labels:<labels:<name:"app.gazette.dev/managed-by" value:"estuary.dev/flow" > labels:<name:"content-type" value:"application/x-ndjson" > labels:<name:"estuary.dev/build" value:"7777777777777777" > labels:<name:"estuary.dev/collection" value:"a/derivation" > > fragment:<length:536870912 compression_codec:GZIP stores:"gs://example-bucket/" refresh_interval:<seconds:300 > flush_interval:<seconds:86400 > path_postfix_template:"utc_date={{.Spool.FirstAppendTime.Format \"2006-01-02\"}}/utc_hour={{.Spool.FirstAppendTime.Format \"15\"}}" > flags:4 max_append_rate:4194304 > > partition_selector:<include:<labels:<name:"estuary.dev/collection" value:"a/derivation" > > exclude:<> > field_selection:<keys:"a_key" values:"flow_published_at" document:"flow_document" > journal_read_suffix:"materialize/example/materialization/a_table" state_key:"a_table" > shard_template:<id:"materialize/example/materialization/ffffffffffffffff" recovery_log_prefix:"recovery" hint_prefix:"/estuary/flow/hints" hint_backups:2 max_txn_duration:<seconds:1200 > labels:<labels:<name:"app.gazette.dev/managed-by" value:"estuary.dev/flow" > labels:<name:"estuary.dev/build" value:"7777777777777777" > labels:<name:"estuary.dev/expose-port" value:"8001" > labels:<name:"estuary.dev/hostname" value:"16480cc2f6ecc2c0" > labels:<name:"estuary.dev/log-level" value:"info" > labels:<name:"estuary.dev/task-name" value:"example/materialization" > labels:<name:"estuary.dev/task-type" value:"materialization" > > ring_buffer_size:65536 read_channel_size:4096 > recovery_log_template:<name:"recovery/materialize/example/materialization/ffffffffffffffff" replication:3 labels:<labels:<name:"app.gazette.dev/managed-by" value:"estuary.dev/flow" > labels:<name:"content-type" value:"application/x-gazette-recoverylog" > labels:<name:"estuary.dev/build" value:"7777777777777777" > labels:<name:"estuary.dev/task-name" value:"example/materialization" > labels:<name:"estuary.dev/task-type" value:"materialization" > > fragment:<length:268435456 compression_codec:SNAPPY stores:"gs://example-bucket/" refresh_interval:<seconds:300 > flush_interval:<seconds:172800 > > flags:4 max_append_rate:4194304 > network_ports:<number:8001 > )
+(*flow.MaterializationSpec)(name:"example/materialization" connector_type:IMAGE config_json:"{\"image\":\"ghcr.io/estuary/materialize-sqlite:dev\",\"config\":{}}" bindings:<resource_config_json:"{\"table\":\"a_table\"}" resource_path:"a_table" collection:<name:"a/derivation" write_schema_json:"{\"$id\":\"file:///build.flow.yaml?ptr=/collections/a~1derivation/schema\",\"properties\":{\"a_key\":{\"type\":\"string\"}},\"required\":[\"a_key\"],\"type\":\"object\"}" key:"/a_key" uuid_ptr:"/_meta/uuid" projections:<ptr:"/_meta/flow_truncated" field:"_meta/flow_truncated" inference:<types:"boolean" title:"Flow truncation indicator" description:"Indicates whether any of the materialized values for this row have been truncated to make them fit inside the limitations of the destination system." exists:MUST > > projections:<ptr:"/a_key" field:"a_key" is_primary_key:true inference:<types:"string" string:<> exists:MUST > > projections:<field:"flow_document" inference:<types:"object" exists:MUST > > projections:<ptr:"/_meta/uuid" field:"flow_published_at" inference:<types:"string" string:<format:"date-time" content_encoding:"uuid" > title:"Flow Publication Time" description:"Flow publication date-time of this document" exists:MUST > > ack_template_json:"{\"_meta\":{\"ack\":true,\"uuid\":\"DocUUIDPlaceholder-329Bb50aa48EAa9ef\"}}" partition_template:<name:"a/derivation/ffffffffffffffff" replication:3 labels:<labels:<name:"app.gazette.dev/managed-by" value:"estuary.dev/flow" > labels:<name:"content-type" value:"application/x-ndjson" > labels:<name:"estuary.dev/build" value:"7777777777777777" > labels:<name:"estuary.dev/collection" value:"a/derivation" > > fragment:<length:536870912 compression_codec:GZIP stores:"gs://example-bucket/" refresh_interval:<seconds:300 > flush_interval:<seconds:86400 > path_postfix_template:"utc_date={{.Spool.FirstAppendTime.Format \"2006-01-02\"}}/utc_hour={{.Spool.FirstAppendTime.Format \"15\"}}" > flags:4 max_append_rate:4194304 > > partition_selector:<include:<labels:<name:"estuary.dev/collection" value:"a/derivation" > > exclude:<> > field_selection:<keys:"a_key" values:"flow_published_at" document:"flow_document" > journal_read_suffix:"materialize/example/materialization/a_table" state_key:"a_table" > shard_template:<id:"materialize/example/materialization/ffffffffffffffff" recovery_log_prefix:"recovery" hint_prefix:"/estuary/flow/hints" hint_backups:2 max_txn_duration:<seconds:1200 > labels:<labels:<name:"app.gazette.dev/managed-by" value:"estuary.dev/flow" > labels:<name:"estuary.dev/build" value:"7777777777777777" > labels:<name:"estuary.dev/log-level" value:"info" > labels:<name:"estuary.dev/task-name" value:"example/materialization" > labels:<name:"estuary.dev/task-type" value:"materialization" > > ring_buffer_size:65536 read_channel_size:4096 > recovery_log_template:<name:"recovery/materialize/example/materialization/ffffffffffffffff" replication:3 labels:<labels:<name:"app.gazette.dev/managed-by" value:"estuary.dev/flow" > labels:<name:"content-type" value:"application/x-gazette-recoverylog" > labels:<name:"estuary.dev/build" value:"7777777777777777" > labels:<name:"estuary.dev/task-name" value:"example/materialization" > labels:<name:"estuary.dev/task-type" value:"materialization" > > fragment:<length:268435456 compression_codec:SNAPPY stores:"gs://example-bucket/" refresh_interval:<seconds:300 > flush_interval:<seconds:172800 > > flags:4 max_append_rate:4194304 > )
